seperate logic into listner hook so that all navigation buttons work on all screens.. [New Session, Toggle Full Screen, View History]
